+++ src/main/java/edu/kis/vh/nursery/stackImplementation/Node.java
@@ -1,0 +1,37 @@
+package edu.kis.vh.nursery.stackImplementation;
+
+public class Node {
+
+    public int value;
+    public Node prev, next;
+
+    protected Node(int i) {
+        setValue(i);
+    }
+
+    public int getValue() {
+        return value;
+    }
+
+    public void setValue(int value) {
+        this.value = value;
+    }
+
+    public Node getPrev() {
+        return prev;
+    }
+
+    public void setPrev(Node prev) {
+        this.prev = prev;
+    }
+
+    public Node getNext() {
+        return next;
+    }
+
+    public void setNext(Node next) {
+        this.next = next;
+    }
+}
+
+
